feature: Item Entity 변경에 따른 변경점 대부분 수정 완료. Custom Pagination으로 변경 시작. seller-write-item.html option에 따라 가격과 재고도 변경될 수 있게 개편.

+++ src/main/java/com/jay/shoppingmall/domain/item/temporary/ItemTemporary.java
@@ -29,9 +29,9 @@ public class ItemTemporary extends BaseTimeEntity {
 
     private String description;
 
-    private Integer price;
+    private Long price;
 
-    private Integer salePrice;
+    private Long salePrice;
 
     private Integer stock;
